backend/pyproject.toml
[project]
name = "ugcgen-backend"
version = "0.1.0"
description = "UGC Video Generator - FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "google-genai>=1.0.0",
    "google-cloud-texttospeech>=2.21.0",
    "google-cloud-storage>=2.18.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    "pillow>=10.0.0",
    "supabase>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "anyio>=4.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
